feat(layout): combine sidebar and footer
