+++ src/components/Layout/Layout.module.css
@@ -1,3 +1,8 @@
+.container{
+
+}
+
+
 .container {
     display: flex;
     justify-content: center;
